add already seen tx metrics to both mempools to monitor duplication in message passing (#875)
This commit adds two metrics AlreadySeenTxs and SuccessfulTxs to both mempool implementations. If there is a high ratio of AlreadySeenTxs to total txs this indicates a high degree of duplication in the messages sent across the wire and validates the hypothesis that a content addressable network would be more advantageous in saving bandwidth (especially given the expected size of a tx).

Co-authored-by: Rootul P <rootulp@gmail.com>