blocksync: run process proposal when verifying new blocks (#929)
Blocksync didn't previously call ProcessProposal. This meant that nodes syncing don't check the Data they receive with the hash in the header. This meant that malicious nodes could fabricate blocks packed with extra unused data or worse they could include transactions that lead to a different app hash causing the syncing node to panic when ApplyBlock errors (because of an app hash mismatch). This commit adds a `ProcessProposal` call in blocksync v0

Co-authored-by: Rootul P <rootulp@gmail.com>