feat: add time to `RequestPreparePropsosal` (#1081)
## Description

Adds the last commit median time to `RequestPreparePropsosal`. This
allows the application to accurately calculate the spendable funds of
vesting accounts, since this is the time used when executing the
transactions and when evaluating the propsoal during `ProcessProposal`.

The height was also added, but it doesn't have to be since the
application can calculate that value. That value is also passed in other
versions of comet, and we'll have to add it we want to unfork from the
sdk so I figured why not. Happy to remove it as well.