chore: backport pyroscope from v0.34.x (#1104)
## Description

backports the pyroscope PR to main. 

note that relative to v0.34.x, we had to upgrade to v1.0.3 to use a new
version of go, and there was a short paragraph of docs added to
hopefully help get started using it.

Let me know @cmwaters if you're running into any other issues that need
to be added to the docs. When I used it it was pretty straightforward.
Just toss the url in the config and set false = true

---------

Co-authored-by: Rootul P <rootulp@gmail.com>
Co-authored-by: Callum Waters <cmwaters19@gmail.com>