chore: Consolidating default SendRate and RecvRate settings for clarity and consistency (#1118)
There are two default SendRate/RecvRate settings in the codebase:
- The first one is a constant defined in the `connection.go` file, which
is set to 512KB/s. You can find it in the code at this
[link](https://github.com/celestiaorg/celestia-core/blob/main/p2p/conn/connection.go#L43).
- The second default value is set in the `config.go` file, which
configures various P2P settings. Among these configurations, the
SendRate and RecvRate default to 5MB/s. You can locate this in the code
at this
[link](https://github.com/celestiaorg/celestia-core/blob/main/config/config.go#L609).

When a node is running, the SendRate and RecvRate values that take
effect are the ones specified in the `config.go` file, which is 5MB/s.

This PR aims to consolidate these two scattered default values. By doing
so, it ensures that when developers consult the code, there will be no
confusion regarding the default settings for these parameters.