fix!: change the instrumentation config to circumvent viper bug (#1143)
## Description

There's an viper property or potential bug that stops us from overriding
the default values for slices of strings. The "solution" atm is to use a
string and parse it later like animals later. Unfortunately I think this
is breaking since we're changing what a config is doing so tbh I'm not
sure if we can ever include this in v1. If that's the case then we might
want to consider not merging this to v0.34.x-celestia and only to main.

closes #1141

---------

Co-authored-by: Rootul P <rootulp@gmail.com>