feat(mempool): add metric size of pool in bytes (#1195)
This PR exposes a new metric SizeBytes for the total size of the mempool
in bytes. We already keep track of this value internally, in the
variable txsBytes. Currently there is the metric Size for the size of
the mempool in number of transactions.

(cherry picked from commit 9f641e7ff15a73e7e6fa520fe15ce6288258adfc)