feat: enhances S3Download to filter by traced table names (#1374)
This PR updates the `S3Download` implementation to allow specifying the
names of the traced tables when downloading from an S3 bucket. This
enhancement enables file filtering and reduces the amount of data
downloaded. It is a desirable feature when interacting with the traced
data of large network tests. If no table name is provided, then
`S3Download` downloads all the traced tables.