Adds logs for the state of the trace pull server (#1401)
As part of migrating to Knuu for the celestia-app e2e tests, I debugged
the tracing feature which required adding log messages to verify that
the tracer pull server is up and running. This PR includes those log
messages, which previously lived in a custom branch and were not
available to everyone.