fix: link for message_block_layout.md (#1559)
**Description of changes**:
Fixed an issue with the file path structure: the first link to the
`message_block_layout.md` file in the `src/rationale` directory was not
accessible because the file should be located directly in the
`rationale` folder. The link has now been corrected to point to the
proper file location.