fix: mempool locking mechanism in v1 and cat (#1582)
This is residual of
https://github.com/celestiaorg/celestia-core/pull/1553

The problem is now even more subtle. Because the mempool mutexes weren't
over both CheckTx and the actual adding of the transaction to the
mempool we occasionally hit a situation as follows:

- CheckTx a tx with nonce 2
- Before saving it to the store, collect all transactions and recheck
tx. This excludes the last tx with nonce 2, thus the nonce in the state
machine is still 1
- Save the tx to the pool
- New tx comes in with nonce 3. The application is at 1 so rejects it
expecting the next to be 2.

This PR fixes this pattern, however this won't be watertight for the CAT
pool until we can order both on gas fee (priority) and nonce.

---------

Co-authored-by: Rootul Patel <rootulp@gmail.com>