chore: change "Failed to load block meta" catchup log to debug (#2122)
main version of #2121 

the node operator shouldn't need to do anything here and this log occurs
every single time they try to send them a block part... so seems like a
bad thing to error on.

to clarify further this log only gets hit when a peer is catching up
from a block height the host doesn't have due to pruning or syncing.