fix(ci): backport rules (#2155)
Mergify hasn't been backporting any PRs to v0.38.x-celestia. The import
issue is that the labels in the config didn't match the labels on the
repo.

While I was here, I removed the merge queue which we don't use, some
unnecessary conditions, a backport section for v0.37.x which we don't
maintain

Reviewers should verify the labels in this PR match the labels on the
repo at https://github.com/celestiaorg/celestia-core/labels?q=backport

![Screenshot 2025-07-07 at 10 09
52 PM](https://github.com/user-attachments/assets/a019dc87-e6ff-42aa-b49c-dd5f98b7a415)