fix: make json decoding old v0.34 EventAttribute backwards compatible (#2184)
makes json decoding the v0.34 proto messages backwards compatible with
the new ones.

the main issue is that while the data over the wire is the same for
bytes and strings in protobuf, the marhalling and unmarshalling of json
is not the same.

The fix here is to decode and encode in a backwards compatible way.

---------

Co-authored-by: Rootul P <rootulp@gmail.com>
(cherry picked from commit bb839a2c5c1ace903f29030f8299661e8dcf646b)