feat: make the propagation reactor backwards compatible (#2149)
closes #1635 

this PR updates the consensus peer state as the propagation reactor
gossips block parts. This enables us to actually test the new reactor as
it is solely responsible for distributing parts among peers that are
running it (assuming the proposer is also creating and distributing a
compact block), while also keeping the old routine running.

Since the old routine is running for all peers, legacy proposal blocks
are distributed exclusively using the legacy routine.

Lastly, it adds the ability to shut off the old routine entirely when
the app version is v5, while also adding a config/flag that can force
the usage of the old routine if something goes wrong with the new one.
After the v5 activation and the network operates normally for a while,
we can remove the old routine entirely.

---------

Co-authored-by: CHAMI Rachid <chamirachid1@gmail.com>