fix: initialize peer state when proposing (#2245)
part of #2240

initializes the peer state before the proposer attempts to update the
peer state for the receiving node. There's not a lot of reason to do
this, since the proposer already has all the block data. That's updated
as comment in this PR as well.

It's also unclear to me in the consensus peer state is getting updated,
which means the redundant part could still be sent.