fix: make the findNewPeerToRequestTx iterative (#2288)
Closes https://github.com/celestiaorg/celestia-core/issues/2271

The underlying cause seems to be `nil` peers not cleaned up. This also
happens in the propagation reactor and we still didn't get to the root
of the issue.

So, I created this PR to avoid the infinite recursions. Then, i'll open
another issue to investigate the nil peer issue.