fix!: stop panics in parallel reactor processing (#2387)
this stops some panics from occurring since the new queues were added on
main.

I might have gone a bit overboard with the tests but given the
importance here I think its worth it.

The only reason for the ! is that I cleaned up a function that returned
an error when it didn't need to, and the only way to handle that error
was to panic