feat: prioritise sending parity chunks when proposing the block (#2413)
Closes https://github.com/celestiaorg/celestia-core/issues/1662

This PR only sends parity shares when proposing the block. However, when
a node decodes the block, it will send haves for all the block. I opted
to leave the second part sending original parts because that's already
~4s in the round, if a node didn't get all the parity parts + mempool
parts to decode, it's better to have more choices of peer to download
from.