fix: limit the number of SeenTx broadcast for new txs (#2522)
One thing that I noticed during our recent traces was that often the
original node operator was uploading the new transactions too much,
which ended up slowing distribution.

We can reduce this and try to force the higher throughput case by doing
the same thing we do when we normally broadcast a seenTx, which is to
limit the number of peers that we send them to.

Here's a trace of the problematic case, where under congestion the
broadcaster is uploading the transaction too much
<img width="1436" height="1069" alt="Screenshot_20251001_155639-1"
src="https://github.com/user-attachments/assets/9015321c-2187-4cef-abf9-d0b757d9ed20"
/>

We want more cases that look like this trace, where we peers of the
broadcaster are doing a lot of work.
<img width="1433" height="849" alt="Screenshot_20251001_155506-1"
src="https://github.com/user-attachments/assets/5784ed52-a0d1-498e-9180-383dcc3f99b6"
/>

for posterity, I'll also link the 
[graph.py](https://github.com/user-attachments/files/22649363/graph.py)
python script to generate these traces here.