feat: account for KMS signing delay in the delayed precommit timeout (#2541)
A second option to close:
https://github.com/celestiaorg/celestia-core/issues/2538

First option: https://github.com/celestiaorg/celestia-core/pull/2540

I couldn't test this in a realistic network, given how hard it is to
simulate a realistic environment with KMS. However, if we agree on this
approach, we can ask POPS to use it and see if this gives them enough
time to include their KMS signatures.

The potential issue with this approach is not knowing exactly how much
latency to expect. For example, if validators are running geographically
distributed KMS, it might take up to a second if we have 5 signers on
different continents.
The other issue is if most validators are using KMS, then they will all
share pre-commit signatures pre-maturely and the network would advance
at 5s blocks. And given most of them will fire faster and the network
advances at a faster block time, it means that we will go back to the
original problem and have validators not able to include their own
signatures in time.

So, this approach depends on the validators running KMS to be low.