fix: receiving have on catchup block due to congestion (#2546)
Closes: https://github.com/celestiaorg/celestia-core/issues/2211

This is issue happens because we receive the votes using the consensus
queues. Then we use those votes to add commitments for catchup blocks.

So, if the node is congested, and it received a `have` for a certain
block, but that `have` wasn't processed in time until a commitment was
added, then this `have` was causing that error.