feat: optionally skip processProposal while blocksyncing (#2658)
first part of #2657

adds a config and flag to the blocksync config that enables a user to
select if they call ProcessProposal while blocksyncing or not. This can
significantly increase syncing rate, especially with large blocks. When
state syncing, users also suffer from a minimal security decrease.

when verifying is disabled, we are only comparing the partsetheader, not
the data root.