fix: eviction due to firstHeight not being updated on removeTx (#2682)
# Overview

If the client submits a lot of txs, then there is a possibility that
`txSet` will never be empty. There was a bug that we never updated
`firstHeight` when the tx was removed, thus we could experience
evictions, because the `txSet` did not reflect the actual lowest height
within the set.

**Steps to reproduce before fix**

Run 50 validators on talis, use the following command (replace the
folder and ip with your ones):
```go run ./tools/latency-monitor -k ./exp/exp-cat-fix/payload/validator-1 -b 104800 -z 104800 -d 1ms -e 157.230.255.154:9091```